refactor image IO to new source file
